refactor: Cars를 별도 클래스로 분리 - GameSetter에서 Cars를 분리 - 관련 테스트 코드도 함께 수정
