+++ src/test/java/racingcar/CarsTest.java
@@ -1,0 +1,39 @@
+package racingcar;
+
+import java.util.ArrayList;
+import java.util.List;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+public class CarsTest {
+
+    @Test
+    void setCarsTest_정상입력() {
+        List<Car> expected = new ArrayList<>();
+        expected.add(new Car("루트", 0));
+        expected.add(new Car("소롱", 0));
+
+        List<Car> actual = new Cars("루트,소롱").getCars();
+        assertThat(actual).usingRecursiveComparison().isEqualTo(expected);
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"한대", "자동차#;"})
+    void setCarsTest_자동차_입력_수(String input) {
+        assertThatThrownBy(() ->
+                new Cars(input)).isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("자동차는 두 대 이상 입력해야 합니다.");
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"루트,소롱,루트", "루트,루트,루트"})
+    void setCarsTest_중복되는_자동차_입력(String input) {
+        assertThatThrownBy(() ->
+                new Cars(input)).isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("중복되는 이름을 입력할 수 없습니다.");
+    }
+}
